backend/requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
psycopg2-binary>=2.9
pymongo>=4.6
django-redis>=5.4
Pillow>=10.0
gunicorn>=22.0
scikit-learn>=1.4
numpy>=1.26
python-decouple>=3.8
